fix: 修复代码审查 9 项发现
#1 [致命] main.py: 还原重构中缺失的 SECRET_KEY, ALGORITHM 导入
#2 [高] fix.py: os.chdir 替换为 git -C + subprocess cwd 参数
#3 [高] reporting.py: CVSS 3.1 公式修正 + round() 改为 math.ceil()
#4 [高] orchestrator.py: _persist_session 包裹在 asyncio.to_thread 中
#5 [中] fix.py: 移除 git add . 回退，file_path 为空时报错
#6 [中] main.py: agent_findings.cvss_score INT → DECIMAL(3,1)
#7 [中] orchestrator.py: pymysql 连接添加 try/finally 保护
#8 [低] fix.py: 分支名用 re.sub 清洗，仅保留安全字符
#9 [低] tests: TestSessionManagement 添加 setup/teardown 隔离

+++ backend/main.py
@@ -64,7 +64,7 @@ if _mail_config_ready():
 
 # ================= 认证模块 (from core.auth) =================
 from core.auth.password import verify_password, get_password_hash
-from core.auth.jwt import create_access_token, get_current_user, oauth2_scheme
+from core.auth.jwt import create_access_token, get_current_user, oauth2_scheme, SECRET_KEY, ALGORITHM
 from core.auth.dependencies import (
     get_current_active_user,
     get_current_admin_user,
@@ -257,7 +257,7 @@ def init_db():
                     description TEXT NULL,
                     evidence JSON NULL,
                     cve_id VARCHAR(50) NULL,
-                    cvss_score INT NULL,
+                    cvss_score DECIMAL(3,1) NULL,
                     file_path VARCHAR(500) NULL,
                     line_number INT NULL,
                     fixed_code TEXT NULL,
